The httpsRequestTo function moved to jokeService.
